altered user,Adminuser and NormalUser classes to have better inheritance
