+++ src/main/java/nl/miwn/ch16/buggies/buggyrecepten/model/AdminUser.java
@@ -2,11 +2,8 @@ package nl.miwn.ch16.buggies.buggyrecepten.model;
 
 import jakarta.persistence.DiscriminatorValue;
 import jakarta.persistence.Entity;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 
-import java.util.Collection;
-import java.util.List;
+import java.util.Set;
 
 /**
 *@author Marnix Ripke
@@ -17,33 +14,8 @@ import java.util.List;
 @DiscriminatorValue("ADMIN")
 public class AdminUser extends User{
 
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        return List.of(new SimpleGrantedAuthority("ROLE_ADMIN"));
+    public AdminUser() {
+        this.setRoles(Set.of("ROLE_ADMIN"));
     }
 
-    @Override
-    public String getUsername() {
-        return name;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return super.isAccountNonExpired();
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return super.isAccountNonLocked();
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return super.isCredentialsNonExpired();
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return super.isEnabled();
-    }
 }
